# docs/swagger/swagger.yaml
definitions:
...
  http.jiraInstanceRequestBody:
    properties:
      access_token:
        type: string
      client_mail:
        type: string
      host:
        type: string
    required:
    - access_token
    - client_mail
    - host
    type: object
  http.jiraStoryJQLSearchRequestBody:
    properties:
      jql:
        type: string
      maxResults:
        type: integer
      startAt:
        type: integer
    required:
    - jql
    type: object
...
  http.standardJsonResponse:
    properties:
      data:
        type: object
      error:
        type: string
      meta:
        type: object
      success:
        type: boolean
    type: object
...
  thunderdome.JiraInstance:
    properties:
      access_token:
        type: string
      client_mail:
        type: string
      created_date:
        type: string
      host:
        type: string
      id:
        type: string
      updated_date:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact:
    name: Steven Weathers
    url: https://github.com/StevenWeathers/thunderdome-planning-poker
  description: |-
    Thunderdome Planning Poker API for both Internal and External use.
    WARNING: Currently not considered stable and is subject to change until 1.0 is released.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Thunderdome API
  version: BETA
paths:
...
  /users/{userId}/jira-instances:
    get:
      description: get list of Jira instances associated to user
      parameters:
      - description: the user ID to find jira instances for
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/thunderdome.JiraInstance'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Get User Jira Instances
      tags:
      - jira
    post:
      description: Creates a Jira Instance associated to user
      parameters:
      - description: the user ID to associate jira instance to
        in: path
        name: userId
        required: true
        type: string
      - description: new jira_instance object
        in: body
        name: jira
        required: true
        schema:
          $ref: '#/definitions/http.jiraInstanceRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: returns new jira instance
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.JiraInstance'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Jira Instance
      tags:
      - jira
  /users/{userId}/jira-instances/{instanceId}:
    delete:
      description: Deletes a Jira Instance associated to user
      parameters:
      - description: the user ID jira instance associated to
        in: path
        name: userId
        required: true
        type: string
      - description: the jira_instance ID to delete
        in: path
        name: instanceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Jira Instance
      tags:
      - jira
    put:
      description: Updates a Jira Instance associated to user
      parameters:
      - description: the user ID jira instance associated to
        in: path
        name: userId
        required: true
        type: string
      - description: the jira_instance ID to update
        in: path
        name: instanceId
        required: true
        type: string
      - description: updated jira_instance object
        in: body
        name: jira
        required: true
        schema:
          $ref: '#/definitions/http.jiraInstanceRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: returns updated jira instance
          schema:
            allOf:
            - $ref: '#/definitions/http.standardJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/thunderdome.JiraInstance'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Jira Instance
      tags:
      - jira
  /users/{userId}/jira-instances/{instanceId}/jql-story-search:
    post:
      description: Queries Jira Instance API for Stories by JQL
      parameters:
      - description: the user ID associated to jira instance
        in: path
        name: userId
        required: true
        type: string
      - description: the jira_instance ID to query
        in: path
        name: instanceId
        required: true
        type: string
      - description: jql search request
        in: body
        name: jira
        required: true
        schema:
          $ref: '#/definitions/http.jiraStoryJQLSearchRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.standardJsonResponse'
      security:
      - ApiKeyAuth: []
      summary: Query Jira for Stories by JQL
      tags:
      - jira
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"